requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pyjwt==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
pydantic[email]
groq==0.11.0
qdrant-client==1.11.3
sentence-transformers==3.2.1
pypdf==4.3.1
langchain-text-splitters==0.3.0
deep-translator==1.11.4
langdetect==1.0.9
redis==5.1.1
fastapi-limiter==0.1.6
aiofiles==24.1.0